trading/marcone-trading-bot-python/py/requirements.txt
python-deriv-api==0.1.6
python-dotenv==1.0.1
reactivex==4.0.4
typing_extensions==4.12.0
websockets==10.3
